Update readme.txt
Hi @dennisnissle ,

i have added the missing informations from the other readme file, changed the Headlines to the readme standard and controlled the URLs with Markdown. After that i tried to use the readme validator, but it didnt work and didnt gave me any result. 
https://wordpress.org/plugins/developers/readme-validator/

So would be nice to check, whether i missunderstand anything and need to rework.

Greetings